European Bank Business Case
Problem statement:
The data is about telemarketing campaigns of a European banking institution.  The European bank wants to predict which clients will secure a term deposit based on a set of information on client and purchase of term deposit. The marketing is usually based on phone calls. Often, a client need to be persuaded multiple times in order to assess if the product (bank term deposit) would be or not subscribed. Predictive modelling approach will help the bank to manage their telemarketing campaign efficiently.
Data Description:
Information was collected on 41,188 clients against 20 variables for the prediction term deposit (yes/no).